+++ data/all_calories.csv
@@ -1,0 +1,47 @@
+date,calory,calory_bmr,calory_activity,calory_out,calory_in
+2019-12-01,2232,1966,323,2289,2114
+2019-12-02,2659,1960,976,2936,793
+2019-12-03,2395,1946,606,2552,1123
+2019-12-04,2474,1954,680,2634,1333
+2019-12-05,2779,1952,1062,3014,1074
+2019-12-06,2291,1952,463,2415,1443
+2019-12-07,2577,1946,817,2763,2755
+2019-12-08,2092,1944,172,2116,1390
+2019-12-09,3138,1946,1557,3503,1437
+2019-12-10,4686,1934,3177,5111,1641
+2019-12-11,3887,1938,2171,4109,1565
+2019-12-12,4002,1942,2363,4305,1726
+2019-12-13,5470,1932,3978,5910,1439
+2019-12-14,3021,1938,1177,3115,1464
+2019-12-15,5337,1932,3931,5863,1354
+2019-12-16,4504,1938,2938,4876,1187
+2019-12-17,5478,1938,4069,6007,0
+2019-12-18,6174,1936,4975,6911,0
+2019-12-19,5542,1935,4256,6191,0
+2019-12-20,3282,1934,1446,3380,0
+2019-12-21,2538,1954,687,2641,0
+2019-12-22,2271,1956,330,2286,0
+2019-12-23,1944,1944,0,1944,0
+2019-12-24,4778,1932,3197,5129,0
+2019-12-25,5533,1930,4200,6130,0
+2019-12-26,4606,1922,3090,5012,0
+2019-12-27,2300,1928,281,2209,0
+2019-12-28,2253,1920,316,2236,0
+2019-12-29,6045,1912,4783,6695,0
+2019-12-30,4248,1915,2673,4588,0
+2019-12-31,3103,1918,1385,3303,0
+2020-01-01,1920,1920,0,1920,0
+2020-01-02,1921,1921,0,1921,0
+2020-01-03,1922,1922,0,1922,0
+2020-01-04,1915,1915,0,1915,0
+2020-01-05,4772,1913,3373,5286,0
+2020-01-06,4519,1911,3097,5008,0
+2020-01-07,3440,1908,1767,3675,0
+2020-01-08,1906,1906,0,1906,0
+2020-01-09,4567,1904,3132,5036,0
+2020-01-10,4346,1894,2915,4809,0
+2020-01-11,4347,1886,2970,4856,1057
+2020-01-12,2131,1886,272,2158,0
+2020-01-13,5259,1886,3947,5833,1193
+2020-01-14,5185,1878,3968,5846,1125
+2020-01-15,2180,1886,280,2166,1387
